+++ R/hp.R
@@ -40,6 +40,7 @@ ssd_hp <- function(x, ...) {
     return(as_tibble(data.frame(
       conc = conc, est = est * 100,
       se = na, lcl = na, ucl = na, dist = rep(dist, length(conc)),
+      wt = rep(1, length(conc)),
       stringsAsFactors = FALSE
     )))
   }
@@ -48,26 +49,30 @@ ssd_hp <- function(x, ...) {
   as_tibble(data.frame(
     conc = conc, est = est * 100,
     se = cis$se * 100, lcl = cis$lcl * 100, ucl = cis$ucl * 100,
-    dist = dist, stringsAsFactors = FALSE
+    dist = dist, wt = 1, stringsAsFactors = FALSE
   ))
 }
 
 .ssd_hp_fitdists <- function(x, conc, ci, level, nboot, parallel, ncpus,
-                             average, ic) {
+                             average) {
   if (!length(x) || !length(conc)) {
     no <- numeric(0)
     return(as_tibble(data.frame(
       conc = no, est = no, se = no,
       lcl = no, ucl = no, dist = character(0),
+      wt = numeric(0),
       stringsAsFactors = FALSE
     )))
   }
 
+  weight <- .ssd_gof_fitdists(x)$weight
   hp <- lapply(x, ssd_hp,
     conc = conc, ci = ci, level = level, nboot = nboot,
     parallel = parallel, ncpus = ncpus
   )
   if (!average) {
+    hp <- mapply(function(x, y) {x$wt <- y; x}, hp, weight, USE.NAMES = FALSE,
+                 SIMPLIFY = FALSE)
     hp <- do.call("rbind", hp)
     row.names(hp) <- NULL
     return(as_tibble(hp))
@@ -77,11 +82,11 @@ ssd_hp <- function(x, ...) {
   hp <- Reduce(function(x, y) {
     abind(x, y, along = 3)
   }, hp)
-  weight <- .ssd_gof_fitdists(x)$weight
   suppressMessages(hp <- apply(hp, c(1, 2), weighted.mean, w = weight))
   hp <- as.data.frame(hp)
   hp$conc <- conc
   hp$dist <- "average"
+  hp$wt <- 1
   as_tibble(hp)
 }
 
@@ -121,14 +126,17 @@ ssd_hp.fitdistcens <- function(x, conc, ci = FALSE, level = 0.95, nboot = 1000,
 ssd_hp.fitdists <- function(x, conc, ci = FALSE, level = 0.95, nboot = 1000,
                             parallel = NULL, ncpus = 1,
                             average = TRUE, ic = "aicc", ...) {
-  chk_string(ic)
-  chk_subset(ic, c("aic", "aicc", "bic"))
   chk_unused(...)
+  
+  if(!missing(ic)) {
+    deprecate_warn("0.3.6", "ssdtools::ssd_hp(ic = )",
+                   details = "AICc is used for model averaging unless the data are censored in which case AIC is used.")
+  }
 
   .ssd_hp_fitdists(x, conc,
     ci = ci, level = level, nboot = nboot,
     parallel = parallel, ncpus = ncpus,
-    average = average, ic = ic
+    average = average
   )
 }
 
@@ -139,13 +147,15 @@ ssd_hp.fitdists <- function(x, conc, ci = FALSE, level = 0.95, nboot = 1000,
 ssd_hp.fitdistscens <- function(x, conc, ci = FALSE, level = 0.95, nboot = 1000,
                                 parallel = NULL, ncpus = 1,
                                 average = TRUE, ic = "aic", ...) {
-  chk_string(ic)
-  chk_subset(ic, c("aic", "bic"))
   chk_unused(...)
+  
+  if(!missing(ic)) {
+    deprecate_warn("0.3.6", "ssdtools::ssd_hp(ic = )",
+                   details = "AICc is used for model averaging unless the data are censored in which case AIC is used.")
+  }
 
   .ssd_hp_fitdists(x, conc,
     ci = ci, level = level, nboot = nboot,
-    parallel = parallel, ncpus = ncpus, average = average,
-    ic = ic
+    parallel = parallel, ncpus = ncpus, average = average
   )
 }