+++ R/namespace.R
@@ -17,6 +17,7 @@
 #' @importFrom graphics plot
 #' @importFrom VGAM vglm
 #' @importFrom tibble as_tibble
+#' @importFrom utils capture.output
 #' @importFrom Rcpp sourceCpp
 #' @useDynLib ssdtools
 NULL