+++ NEWS.md
@@ -1,9 +1,10 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
-# ssdtools 0.3.5.9000
+# ssdtools 0.3.6
 
-- Same as previous version.
-
+- Added `wt` (Akaike weight) column to `predict()`, `ssd_hc()` and `ssd_hp()`
+- Deprecated argument `ic` to `predict()`, `ssd_hc()` and `ssd_hp()` because unused.
+- Silenced output from `ssd_fit_dists()`.
 
 # ssdtools 0.3.5